+++ src/tree/lazyPropagation/SegmentTreeWithLazyPropagation.h
@@ -5,10 +5,11 @@
 using namespace std;
 void buildAddOnRange(vector<long long> &tree, const vector<long long> &arr, int v, int tl, int tr);
 void updateAddOnRange(vector<long long> &tree, int v, int tl, int tr, int l, int r, int add);
-int getAddOnRange (vector<long long> &tree, int v, int tl, int tr, int pos);
+long long getAddOnRange (vector<long long> &tree, int v, int tl, int tr, int pos);
+
 void buildAssignOnRange(vector<long long> &tree, const vector<long long> &arr, int v, int tl, int tr);
 void push (vector<long long> &tree, int v);
 void updateAssignOnRange (vector<long long> &tree, int v, int tl, int tr, int l, int r, int val);
-int getAssignOnRange (vector<long long> &tree, int v, int tl, int tr, int pos);
+long long getAssignOnRange (vector<long long> &tree, int v, int tl, int tr, int pos);
 
 #endif //SEGMENT_TREE_REALIZATION_SEGMENTTREEWITHLAZYPROPAGATION_H
